feat(frontend): replace popup auth with redirect auth
