+++ src/frontend/src/api/Login.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import { getUserDetailsFromApi } from '@/utilfunctions/login';
 import { CommonActions } from '@/store/slices/CommonSlice';
+import { LoginActions } from '@/store/slices/LoginSlice';
 
 export const TemporaryLoginService: Function = (url: string) => {
   return async (dispatch) => {
@@ -8,8 +9,8 @@ export const TemporaryLoginService: Function = (url: string) => {
       // Sets a cookie in the browser that is used for auth
       await axios.get(url);
 
-      const loginSuccess = await getUserDetailsFromApi();
-      if (!loginSuccess) {
+      const apiUser = await getUserDetailsFromApi();
+      if (!apiUser) {
         dispatch(
           CommonActions.SetSnackBar({
             open: true,
@@ -18,7 +19,10 @@ export const TemporaryLoginService: Function = (url: string) => {
             duration: 2000,
           }),
         );
+        return;
       }
+
+      dispatch(LoginActions.setAuthDetails(apiUser));
     };
 
     await getTemporaryLogin(url);
