+++ src/frontend/src/utilities/CustomDrawer.jsx
@@ -5,7 +5,7 @@ import { Modal } from '@/components/common/Modal';
 import CoreModules from '@/shared/CoreModules';
 import AssetModules from '@/shared/AssetModules';
 import { NavLink } from 'react-router-dom';
-import { createLoginWindow, revokeCookie } from '@/utilfunctions/login';
+import { revokeCookie } from '@/utilfunctions/login';
 import { CommonActions } from '@/store/slices/CommonSlice';
 import { LoginActions } from '@/store/slices/LoginSlice';
 import { ProjectActions } from '@/store/slices/ProjectSlice';
@@ -140,12 +140,12 @@ export default function CustomDrawer({ open, placement, size, type, onClose, onS
                 ml={'3%'}
                 spacing={1}
               >
-                {authDetails['picture'] !== 'null' && authDetails['picture'] ? (
+                {authDetails['img_url'] !== 'null' && authDetails['img_url'] ? (
                   <CoreModules.Stack
                     className="fmtm-w-7 fmtm-h-7 fmtm-flex fmtm-items-center fmtm-justify-center fmtm-overflow-hidden fmtm-rounded-full fmtm-border-[1px]"
                     sx={{ display: { xs: 'block', md: 'none' }, mt: '3%' }}
                   >
-                    <img src={authDetails['picture']} alt="Profile Picture" />
+                    <img src={authDetails['img_url']} alt="Profile Picture" />
                   </CoreModules.Stack>
                 ) : (
                   <AssetModules.PersonIcon color="success" sx={{ display: { xs: 'block', md: 'none' }, mt: '1%' }} />
