+++ src/frontend/src/views/OsmAuth.tsx
@@ -1,0 +1,47 @@
+import React, { useEffect, useState } from 'react';
+import { useNavigate, useLocation } from 'react-router-dom';
+import CoreModules from '@/shared/CoreModules.js';
+import { LoginActions } from '@/store/slices/LoginSlice';
+import { getUserDetailsFromApi } from '@/utilfunctions/login';
+
+function OsmAuth() {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const dispatch = CoreModules.useAppDispatch();
+  const [isReadyToRedirect, setIsReadyToRedirect] = useState(false);
+
+  useEffect(() => {
+    // Redirect workaround require for locahost, until PR is merged:
+    // https://github.com/openstreetmap/openstreetmap-website/pull/4287
+    if (window.location.href.includes('127.0.0.1:7051')) {
+      // Pass through same url params
+      window.location.href = `http://fmtm.localhost:7050${location.pathname}${location.search}`;
+      return;
+    }
+
+    const params = new URLSearchParams(location.search);
+    let authCode = params.get('code');
+    let state = params.get('state');
+
+    // authCode is passed from OpenStreetMap redirect, so get cookie, then redirect
+    if (authCode) {
+      const callbackUrl = `${import.meta.env.VITE_API_URL}/auth/callback/?code=${authCode}&state=${state}`;
+
+      const completeLogin = async () => {
+        // NOTE this encapsulates async methods to call sequentially
+        // Sets a cookie in the browser that is used for auth
+        await fetch(callbackUrl, { credentials: 'include' });
+        const apiUser = await getUserDetailsFromApi();
+        dispatch(LoginActions.setAuthDetails(apiUser));
+      };
+      completeLogin();
+    }
+
+    setIsReadyToRedirect(true);
+    navigate('/');
+  }, [dispatch, location.search, navigate]);
+
+  return <>{!isReadyToRedirect ? null : <div>redirecting</div>}</>;
+}
+
+export default OsmAuth;
